+++ backend/src/main/java/ku_rum/backend/domain/building/domain/BuildingAbbrev.java
@@ -1,8 +1,4 @@
-package ku_rum.backend.domain.building.service;
-
-import ku_rum.backend.global.response.status.BaseExceptionResponseStatus;
-
-import java.util.Optional;
+package ku_rum.backend.domain.building.domain;
 
 public enum BuildingAbbrev {
