+++ backend/src/main/java/ku_rum/backend/domain/building/domain/repository/BuildingRepository.java
@@ -1,4 +1,4 @@
-package ku_rum.backend.domain.building.repository;
+package ku_rum.backend.domain.building.domain.repository;
 
 import ku_rum.backend.domain.building.domain.Building;
 import org.springframework.data.jpa.repository.JpaRepository;
