+++ backend/src/main/java/ku_rum/backend/domain/building/presentation/BuildingSearchController.java
@@ -1,7 +1,7 @@
-package ku_rum.backend.domain.building.controller;
+package ku_rum.backend.domain.building.presentation;
 
 import ku_rum.backend.domain.building.dto.response.BuildingResponse;
-import ku_rum.backend.domain.building.service.BuildingSearchService;
+import ku_rum.backend.domain.building.application.BuildingSearchService;
 import ku_rum.backend.domain.category.response.CategoryDetailResponse;
 import ku_rum.backend.global.response.BaseResponse;
 import ku_rum.backend.global.response.status.BaseExceptionResponseStatus;
