+++ backend/src/main/resources/application.yml
@@ -1,0 +1,56 @@
+spring:
+  profiles:
+    default: local
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:mysql://localhost:3306/kurum?useSSL=false&serverTimezone=Asia/Seoul&characterEncoding=UTF-8
+    username: root
+    password: 1234
+
+  jpa:
+    open-in-view: true
+    hibernate:
+      ddl-auto: create
+  data:
+    redis:
+      host: localhost #my-cache-server #redis #localhost #${EC2_public_IPv4_address}
+      port: 6379
+  # 임시
+  mail:
+    host: smtp.example.com           # SMTP 서버 주소 (여기선 예시로 설정)
+    port: 587                        # SMTP 포트 (일반적으로 587 포트는 TLS/STARTTLS 사용)
+    username: your-email@example.com # 이메일 계정
+    password: your-password          # 이메일 계정 비밀번호
+    properties:
+      mail:
+        smtp:
+          auth: true                  # SMTP 인증을 활성화
+          starttls:
+            enable: true              # STARTTLS 활성화
+          connectiontimeout: 5000     # 연결 타임아웃 (밀리초 단위)
+
+springdoc:
+  swagger-ui:
+
+    groups-order: DESC # path, query, body, response 순으로 출력
+
+    tags-sorter: alpha # 태그를 알파벳 순으로 정렬
+
+    operations-sorter: method  # delete - get - patch - post - put 순으로 정렬, alpha를 사용하면 알파벳 순으로 정렬 가능
+
+  paths-to-match:
+    - / # swagger-ui에 표시할 api의 엔드포인트 패턴
+
+server:
+  port: 8080
+
+
+logging:
+  level:
+    org.springframework.cache : trace
+
+---
+spring:
+  config:
+    activate:
+      on-profile : dev
